refactor(mbk/leases): split lease_template_service.py (926 LOC) into CRUD + placeholder + render
- lease_template_placeholder_service.py (new, 182 LOC): MIME constants,
  text-extraction helpers, and AI placeholder suggestion (suggest_ai_placeholders).
  Defines TemplateNotFoundError and StorageNotConfiguredError as the shared
  exception hierarchy so signed_lease_service imports cleanly.

- lease_template_render_service.py (new, 44 LOC): load_template_source_texts
  for the signed-lease render pipeline.

- lease_template_service.py (kept, 789 LOC): all CRUD entry points (upload,
  list, get, generate_defaults, update_metadata, update_placeholder,
  soft_delete, replace_files). Re-imports constants, helpers, and exceptions
  from the sub-modules so every existing caller (api/lease_templates.py,
  api/signed_leases.py, signed_lease_service.py, tests) is unchanged.

No behavioural changes. 55/55 lease template tests pass.
